fix backend fetching
